+++ src/ALife/Creatur/Gene/Test.hs
@@ -28,25 +28,30 @@ module ALife.Creatur.Gene.Test
     prop_diploid_readable,
     prop_show_read_round_trippable,
     prop_makeSimilar_works,
+    prop_diff_can_be_0,
+    prop_diff_can_be_1,
+    prop_diff_is_symmetric,
     randomTestPattern,
     testPatternDiff,
     makeTestPatternSimilar,
     divvy
   ) where
 
-import           ALife.Creatur.Gene.Numeric.UnitInterval
-    (UIDouble, doubleToUI, uiToDouble)
-import           ALife.Creatur.Gene.Numeric.Util
-    (adjustNum, forceToWord8, scaleFromWord8, scaleWord8ToInt)
+import           ALife.Creatur.Gene.Numeric.UnitInterval (UIDouble, narrow,
+                                                          wide)
+import           ALife.Creatur.Gene.Numeric.Util         (forceToWord8,
+                                                          scaleFromWord8,
+                                                          scaleWord8ToInt)
 import qualified ALife.Creatur.Genetics.BRGCWord8        as W8
 import           ALife.Creatur.Genetics.Diploid          (Diploid, express)
 import           ALife.Creatur.Util                      (fromEither)
 import           Control.DeepSeq                         (NFData, deepseq)
-import           Control.Monad.Random
-    (Rand, RandomGen, getRandom)
+import           Control.Monad.Random                    (Rand, RandomGen,
+                                                          getRandom)
 import           Control.Monad.State.Lazy                (runState)
-import           Data.Serialize
-    (Serialize, decode, encode)
+import qualified Data.Datamining.Pattern.Numeric         as N
+import           Data.Serialize                          (Serialize, decode,
+                                                          encode)
 import           Data.Word                               (Word8)
 import           GHC.Generics                            (Generic)
 import           Test.QuickCheck
@@ -67,17 +72,16 @@ arb8BitInt interval = do
 
 -- | Verify that a value is not affected by serialisation and
 --   deserialisation.
-prop_serialize_round_trippable :: (Eq a, Serialize a) => a -> Property
-prop_serialize_round_trippable x = property $ x' == Right x
+prop_serialize_round_trippable :: (Eq a, Serialize a) => a -> Bool
+prop_serialize_round_trippable x = x' == Right x
   where bs = encode x
         x' = decode bs
 
 -- | Verify that a value is not affected by encoding as a gene, then
 --   decoding that gene.
 prop_genetic_round_trippable :: (Eq g, W8.Genetic g, Show g) =>
-  (g -> g -> Bool) -> g -> Property
-prop_genetic_round_trippable eq g = property $
-  g' `eq` g && null leftover
+  (g -> g -> Bool) -> g -> Bool
+prop_genetic_round_trippable eq g = g' `eq` g && null leftover
   where x = W8.write g
         (result, (_, i, _)) = runState W8.get (x, 0, [])
         leftover = drop i x
@@ -92,25 +96,24 @@ prop_genetic_round_trippable2 n xs dummy = length xs >= n
         xs' = W8.write (g `asTypeOf` dummy)
 
 -- | Verify that two identical genes are expressed as that gene.
-prop_diploid_identity :: Diploid g => (g -> g -> Bool) -> g -> Property
-prop_diploid_identity eq g = property $ express g g `eq` g
+prop_diploid_identity :: Diploid g => (g -> g -> Bool) -> g -> Bool
+prop_diploid_identity eq g = express g g `eq` g
 
 -- | Verify a value is not affected by `show`ing and then `read`ing it.
 prop_show_read_round_trippable
-  :: (Read a, Show a) => (a -> a -> Bool) -> a -> Property
-prop_show_read_round_trippable eq x
-  = property $ (read . show $ x) `eq` x
+  :: (Read a, Show a) => (a -> a -> Bool) -> a -> Bool
+prop_show_read_round_trippable eq x = (read . show $ x) `eq` x
 
 -- | Verify that expressing a diploid gene does not cause an error.
 prop_diploid_expressable
-  :: (Diploid g, W8.Genetic g, NFData g) => g -> g -> Property
-prop_diploid_expressable a b = property $ deepseq (express a b) True
+  :: (Diploid g, W8.Genetic g, NFData g) => g -> g -> Bool
+prop_diploid_expressable a b = deepseq (express a b) True
 
 -- | Verify that reading a diploid gene does not cause an error.
 prop_diploid_readable
   :: (Diploid g, W8.Genetic g, NFData g)
-    => g -> g -> Property
-prop_diploid_readable a b = property $ deepseq (c `asTypeOf` a) True
+    => g -> g -> Bool
+prop_diploid_readable a b = deepseq (c `asTypeOf` a) True
   where ga = W8.write a
         gb = W8.write b
         (Right c) = W8.runDiploidReader W8.getAndExpress (ga, gb)
@@ -119,12 +122,21 @@ prop_diploid_readable a b = property $ deepseq (c `asTypeOf` a) True
 --   away from `b` than `a` was.
 prop_makeSimilar_works
   :: (a -> a -> UIDouble) -> (a -> UIDouble -> a -> a) -> a -> UIDouble
-    -> a -> Property
-prop_makeSimilar_works diff makeSimilar x r y
-  = property $ diffAfter <= diffBefore
+    -> a -> Bool
+prop_makeSimilar_works diff makeSimilar x r y = diffAfter <= diffBefore
   where diffBefore = diff x y
         y' = makeSimilar x r y
         diffAfter = diff x y'
+
+prop_diff_can_be_0 :: (a -> a -> UIDouble) -> a -> Bool
+prop_diff_can_be_0 diff x = diff x x == narrow 0
+
+prop_diff_can_be_1 :: Bounded a => (a -> a -> UIDouble) -> a -> Bool
+prop_diff_can_be_1 diff dummy
+  = diff (minBound `asTypeOf` dummy) (maxBound `asTypeOf` dummy) == narrow 1
+
+prop_diff_is_symmetric :: (a -> a -> UIDouble) -> a -> a -> Bool
+prop_diff_is_symmetric diff x y = diff x y == diff y x
 
 -- | A simple pattern that is useful for testing.
 newtype TestPattern = TestPattern Word8
@@ -138,7 +150,7 @@ instance Arbitrary TestPattern where
 --   between 0 (identical) and 1 (maximally different).
 testPatternDiff :: TestPattern -> TestPattern -> UIDouble
 testPatternDiff (TestPattern x) (TestPattern y)
-  = doubleToUI $ abs (fromIntegral x - fromIntegral y) / 255
+  = narrow $ abs (fromIntegral x - fromIntegral y) / 255
 
 -- | @'makeTestPatternSimilar' target r x@ adjusts @x@ to make it
 --   similar to @target@.
@@ -151,7 +163,7 @@ makeTestPatternSimilar (TestPattern target) r (TestPattern x)
     = TestPattern (forceToWord8 x'')
     where t' = fromIntegral target :: Double
           x' = fromIntegral x :: Double
-          x'' = adjustNum t' (uiToDouble r) x'
+          x'' = N.makeSimilar t' (wide r) x'
 
 -- | Random pattern generator.
 randomTestPattern :: RandomGen r => Rand r TestPattern
@@ -163,4 +175,3 @@ randomTestPattern = TestPattern <$> getRandom
 divvy :: Int -> Int -> Gen [Int]
 divvy n k = vectorOf k $ choose (1, n')
   where n' = max 1 (n `div` k)
-