+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for creatur-numeric-genes
 
+1.0.2 Added Narrow and NonNegative modules.
+      The Narrow class makes it easy to create numeric types with
+      constrained ranges. See PlusMinusOne and NonNegative for examples.
+
 1.0.1 De-linting.
 
 1.0.0 Initial release